requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.2
django-appconf==1.1.0
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-crispy-forms==2.3
django-imagekit==5.0.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-taggit==6.1.0
mysqlclient==2.2.7
pilkit==3.0
pillow==11.1.0
python-dotenv==1.1.0
sqlparse==0.5.3
